refactor(renderer): decompose LayoutRenderer and renderUtils into focused modules
Split monolithic LayoutRenderer.tsx (1,515→357 lines) and renderUtils.ts
(383→29 lines barrel) into single-responsibility modules:

- RenderingPipeline.ts (697 lines): SVG rendering traversal
- HighlightController.ts (298 lines): rAF highlight loop + CSS toggling
- LoopOverlayRenderer.ts (143 lines): loop region overlays
- InteractionHandler.ts (45 lines): click event delegation
- svgHelpers.ts (33 lines): SVG element factories
- viewportUtils.ts (133 lines): viewport queries + binary search
- renderConfigUtils.ts (86 lines): config factories + validation

renderUtils.ts preserved as barrel re-export for backward compatibility.
All 1,690 tests pass. No circular dependencies. Zero visual regressions.

Closes #058
